refactor: make beforeSend callback synchronous
The `beforeSend` callback in `CXExporterOptions` and its internal handling in `MethodChannelCxFlutterPlugin` have been changed from asynchronous (`Future<T>`) to synchronous (`T`).

This change simplifies the event processing logic by removing the need for `await` when invoking the callback. The event stream processing on the native side (specifically iOS) is also adjusted to reflect this synchronous nature.