fix: Make NetworkRequestContext.duration non-nullable
The `duration` field in `NetworkRequestContext` has been made non-nullable and defaults to 0.

This change ensures that `duration` always has a value, simplifying downstream processing and preventing potential null pointer exceptions.

The corresponding `fromJson` method in the generated file `cx_types.g.dart` has been updated to reflect this change, providing a default value of 0 if `duration` is null in the JSON input.